Move Carousel to own file, add carousel to reflections, minor refactor
