+++ design-your-life/src/components/Reflections/reflections.jsx
@@ -2,7 +2,6 @@ import React from 'react';
 import styled from 'styled-components';
 
 const reflections = props => {
-  console.log(props);
   return (
     <>
       <ReflectionBlock>
